# docs/contracts/core-service/openapi.yaml
openapi: 3.0.3
info:
  title: core-service — accounts
  description: >
    Internal endpoints for customer profiles, account balances, and channel
    authentication. core-service is the only component with database access;
    these endpoints are internal-network-only and are not exposed through any
    BFF. Every endpoint requires the calling BFF's service credential; the
    seven domain endpoints additionally require a valid user bearer token
    carrying the scope shown in that operation's description. The four
    pre-auth endpoints (called before any user token exists) require only the
    service credential.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
security:
  - ServiceCredential: []
paths:
  /internal/auth/atm/pin-verifications:
    post:
      summary: Verify a card PIN and report the owning customer id
      description: >
        Pre-auth endpoint (service credential only). The PIN is accepted only
        in the request body, never a query parameter or path segment. Served
        exclusively over this operation's TLS-only connector (design.md
        Decision 8); the same request over the plain-HTTP connector is
        rejected. On a third consecutive failure the card locks (423);
        further attempts, including a correct PIN, are rejected identically
        until a developer resets it (scripts/reset-dev-card-lock.sh — there
        is no customer-facing unlock path). An unknown card number is
        rejected identically to an incorrect PIN (401), never distinguished,
        to avoid a card-existence oracle.
      operationId: verifyPin
      servers:
        - url: https://localhost:8453
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PinVerificationRequest'
      responses:
        '200':
          description: Correct PIN; resets the failure count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PinVerificationResponse'
        '401':
          description: Incorrect PIN or unknown card (indistinguishable)
        '423':
          description: Card is locked (third consecutive failure, or already locked)
  /internal/auth/web/refresh-tokens:
    post:
      summary: Rotate a web refresh token
      description: >
        Pre-auth endpoint (service credential only). Rotation is not
        idempotent: presenting an already-used (rotated or revoked) token is
        rejected and revokes the rest of that token's chain, rather than
        replaying the prior result.
      operationId: rotateWebRefreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: A new session and refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    ServiceCredential:
      type: apiKey
      in: header
      name: X-Service-Credential
      description: >
        Static, per-BFF shared secret. Required on every /internal/**
        endpoint, including the four pre-auth ones.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        HS256 JWT identifying the end customer, channel, and scope.
        Required in addition to ServiceCredential on the seven domain
        endpoints; see each operation's description for its required
        scope(s).
...
  responses:
    NotFound:
      description: No resource matches the given identifier
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    PinVerificationRequest:
      type: object
      required: [cardNumber, pin]
      properties:
        cardNumber:
          type: string
        pin:
          type: string
    PinVerificationResponse:
      type: object
      required: [customerId]
      properties:
        customerId:
          type: string
    RefreshTokenRequest:
      type: object
      description: >
        customerId identifies the token's owner for the web flow; the mobile
        flow instead binds ownership to the deviceId path segment and may
        omit it.
      required: [refreshToken]
      properties:
        customerId:
          type: string
          nullable: true
        refreshToken:
          type: string
    RefreshTokenResponse:
      type: object
      required: [customerId, refreshToken, expiry]
      properties:
        customerId:
          type: string
        refreshToken:
          type: string
        expiry:
          type: string
          format: date-time
    ProblemDetail:
      type: object
      description: RFC 7807 problem details
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri